third_party/WebKit/LayoutTests/fast/html/imports/import-custom-element-in-grandchild.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../http/tests/htmlimports/resources/import-custom-element-helper.js"></script>
</head>
<body>
<script>
var Hello = registerTestingCustomElement('x-hello');
var numberOfLinks = 3;
var t1 = async_test('Custom elements in an indirect import are upgraded.');
var t2 = async_test('Custom elements in an indirect import are resolved.');
var latch = new ImportTestLatch(function() {
    t1.step(function() {
        assert_equals('hello-1', Hello.ids[0]);
        assert_array_equals(['hello-1', 'hello-2'], Hello.ids.slice(0).sort());
        t1.done();
    }, 0);

    t2.step(function() {
        var Bye = registerTestingCustomElement('x-bye');
        assert_equals('bye-1', Bye.ids[0]);
        assert_array_equals(['bye-1', 'bye-2'], Bye.ids.slice(0).sort());
        t2.done();
    });
}, numberOfLinks);
</script>
<link rel="import" href="resources/custom-element-hello-parent-12.html" onload="latch.loaded()">
</body>
</html>
